Update jam file.  Move tests to single code.
